Modified features included in fitting/predictions
I had accidentally included drug age and another feature or two that may have led to some multicollinearity (i.e. drug age and the other 'effective_date' features).  I also decided include another feature or two that had originally been excluded, such as `route`, which indicates the method by which the drug is delivered.  Theoretical basis for inclusion is that some methods of delivery may cost more to produce than others, and this difference is reflected in the price of the drug.